+++ src/App.tsx
@@ -1,13 +1,11 @@
-import { BrowserRouter, Routes, Route } from 'react-router-dom'
+import { Routes, Route } from 'react-router-dom'
 import LibraryPage from './pages/LiberyPage'
 
 function App() {
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route index element={<LibraryPage />} />
-      </Routes>
-    </BrowserRouter>
+    <Routes>
+      <Route index element={<LibraryPage />} />
+    </Routes>
   )
 }
